Changed the class names and added methods for the bookstore.
Changed the name of the class BookStrock to Book and ManageStock to
Bookstore.
Added a property of type UUID to the Book class.
Added methods for:
adding a book: addBook();
listing all the books and the number of their available copies
listBooksAndStockCount();
selling a book: sellBook();
removing a book: removeBook();